nearbyposts/requirements.txt
Django==4.0.4
psycopg2==2.8.5
djangorestframework==3.13.1
django-environ==0.4.5